Refactor mctrivers
